src/test/javascript/spec/app/entities/vscommerce/product-brand/product-brand.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { ProductBrandService } from 'app/entities/vscommerce/product-brand/product-brand.service';
import { IProductBrand, ProductBrand } from 'app/shared/model/vscommerce/product-brand.model';

describe('Service Tests', () => {
  describe('ProductBrand Service', () => {
    let injector: TestBed;
    let service: ProductBrandService;
    let httpMock: HttpTestingController;
    let elemDefault: IProductBrand;
    let expectedResult: IProductBrand | IProductBrand[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(ProductBrandService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new ProductBrand(0, 'AAAAAAA', 'AAAAAAA', 0, 'AAAAAAA', 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a ProductBrand', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new ProductBrand()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a ProductBrand', () => {
        const returnedFromService = Object.assign(
          {
            name: 'BBBBBB',
            shortLabel: 'BBBBBB',
            sortOrder: 1,
            iconFont: 'BBBBBB',
            thumbnailUrl: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of ProductBrand', () => {
        const returnedFromService = Object.assign(
          {
            name: 'BBBBBB',
            shortLabel: 'BBBBBB',
            sortOrder: 1,
            iconFont: 'BBBBBB',
            thumbnailUrl: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a ProductBrand', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
